line_provider/pyproject.toml
[tool.poetry]
name = "line_provider"
version = "0.1.0"
description = ""
authors = ["Артём Поташов <artem.potashov@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
tortoise-orm = "^0.20.0"
uvicorn = "^0.25.0"
asyncpg = "^0.29.0"
scheduler = "^0.8.5"
celery = "^5.3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"


[tool.aerich]
tortoise_orm = "tortoise_conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
